Type checking enforced in simulator
Types are now enforced when assigning values to CPFs -- the default method "safe" will raise an exception if the value of the evaluated expression of a CPF returns a type that is not compatible with the CPF definition (e.g. assigning a real value to a bool pvariable). Due to the 64-bit values of the default RNG, the default np types are now also fixed at int64 and float64.

the below changes are made to allow compatibility with existing rddl files:
- some functions like sgn, round, floor etc produce integers as output so they can be assigned to int-valued CPFs
- better message handling of cast check fails
- RecSim domain is fixed